big header file cleanup; geometric.hpp is no more, now split into several smaller files

+++ gl/gl-async-gpu-timer.hpp
@@ -16,7 +16,7 @@ class GlGpuTimer
     };
 
     GLsync sync;
-    GLuint activeIdx;
+    size_t activeIdx;
     std::vector<query_timer> queries;
 
 public:
